+++ src/copy.js
@@ -30,7 +30,7 @@ module.exports = {
     inviteToServer: inviteLink => `You're membership can also be used in this discord community: ${inviteLink}`,
     memberRoleHasBeenRemoved: serverName => `Your \`Member\` role has been removed at the **${serverName}** discord server. Consider re-newing your CougarCS membership here: <https://www.cougarcs.com/register/>`,
     ABOUT: `I was created in the Summer of 2021 by a gentlemen named Adil Iqbal. My purpose is to securely grant the \`Member\` role to Discord users that have a valid & current CougarCS Membership. To see what I can do for you, please use the \`${config.prefix}help\` command. I have also been given a letter for you from my creator. To read it, use the \`${config.prefix}letter\` command.`,
-    LETTER_FROM_ME: 
+    LETTER_FROM_ME_01: 
     `\`\`\`
     Dear reader,
 
@@ -61,6 +61,9 @@ module.exports = {
 
     Yours truly,
     Adil Iqbal
+    \`\`\``,
+    LETTER_FROM_ME_02:
+    `\`\`\`
     
     -- Footnotes --